+++ Changelog.md
@@ -1,5 +1,6 @@
 # CuteCharts 0.0.1 pre-Alpha
 
+- add static vertical and horizontal lines
 - variable autoscale modus
 - add Min/Max Scaling as option to toolbox